requirements.txt
async-generator==1.10
attrs==21.2.0
azure-core==1.19.0
azure-identity==1.7.0
Brotli==1.0.9
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
ConfigArgParse==1.5.3
confluent-kafka==1.7.0
cryptography==35.0.0
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.8.0
geventhttpclient==1.5.3
google-api-core==2.2.0
google-auth==2.3.2
googleapis-common-protos==1.53.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
locust==2.4.1
locust-plugins==1.7.3
lxml==4.6.3
MarkupSafe==2.0.1
msal==1.15.0
msal-extensions==0.3.0
msgpack==1.0.2
opencensus==0.8.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.0
outcome==1.1.0
paho-mqtt==1.6.1
portalocker==1.7.1
protobuf==3.19.0
psutil==5.8.0
psycogreen==1.0.2
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.3.0
pymongo==3.12.1
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pyzmq==22.3.0
requests==2.26.0
roundrobin==0.0.2
rsa==4.7.2
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket-client==1.2.1
Werkzeug==2.0.2
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
